Añadidas excepciones; Cantidades pasadas de int a double

+++ src/main/java/tds/apoyanos/modelo/Apoyo.java
@@ -2,29 +2,26 @@ package tds.apoyanos.modelo;
 
 import java.util.Date;
 
-//TODO Falta la cantidad
-//TODO Constructor que permita no usuario.
-
 public class Apoyo {
 	private String comentario;
 	private Date fecha;
 	private Usuario usuario;
 	private Recompensa recompensa;
-    private int cantidad;
+    private double cantidad;
 
-    public Apoyo(String comentario, Recompensa recompensa, int cantidad){
+    public Apoyo(String comentario, Recompensa recompensa, double cantidad){
         this.comentario = comentario;
         this.recompensa = recompensa;
         this.cantidad = cantidad;
         this.fecha = new Date();
     }
 
-    public Apoyo(Usuario usuario, String comentario, Recompensa recompensa, int cantidad){
+    public Apoyo(Usuario usuario, String comentario, Recompensa recompensa, double cantidad){
 		this(comentario,recompensa, cantidad);
         this.usuario = usuario;
 	}
 
-    public int getCantidad() {
+    public double getCantidad() {
         return cantidad;
     }
